Add Futures compatibility API
This adds an `Instrument` trait which allows futures to be extended with an
`instrument` method, inspired by
[tracing-futures](https://github.com/tokio-rs/tracing/tree/master/tracing-futures).
This allows a span to be correctly associated with a given task, and
fixes the issues highlighted by #43.

Using this API requires moving the span into the future it is
instrumenting, so a `clone_span` method on `Tracer` was added to allow a
given span to be associated with many futures.