Remove tracer name prefix from span name
Currently spans created with a tracer that has a name are augmented to
have their names prepended with the tracer name. This is not then
intended behaviour of a named tracer and needs to be changed.

The expected behaviour is that the span name should not be modified from
the one passed to the SDK.