Add otlp exporter pipeline
This change adds a pipeline builder to `opentelemetry-otlp`. It
simplifies recommended configuration and adds documentation to help
people get started quickly. Simple example:

```rust
use opentelemetry::api::Tracer;

fn main() -> Result<(), Box<dyn std::error::Error>> {
    let tracer = opentelemetry_otlp::new_pipeline().install()?;

    tracer.in_span("doing_work", |cx| {
        // Traced app logic here...
    });

    Ok(())
}
```

A builder style pattern is also added to the sdk trace `Config` to allow
for a consistent configuration in the full kitchen sink example added to
the docs as well.

The `tokio` and `async-std` optional features are moved to
`opentelemetry` from `opentelemetry-jaeger` to avoid duplicating the
logic in `opentelemetry-otlp`. This allows trace provider builders to
have a `with_exporter` that can be a batch exporter or simple exporter
depending on the specified feature.