[metric] Add DDSketchAggregator.
DDSketch is a distributed quantile sketch algorithm which is fully mergeable, meaning that multiple sketches from distributed system can be combined easily.

DDSketch has a relative error guarantee of alpha for any quantile q in [0, 1] that is not too small. That's to say, the q-quantile will be accurate up to a relative error of alpha as long as it belongs to one of the bins within the maximum number of bins.

More detail explanation on how DDSketch works, please refer to http://www.vldb.org/pvldb/vol12/p2195-masson.pdf