Add pipeline uninstall mechanism (#229)
In order to address the issues raised in #225, the following changes are
proposed:

1) Fix the current bug where futures are being dropped instead of
properly awaited by updating:

```diff
- let spawn = |future| tokio::task::spawn_blocking(|| future);
+ let spawn = |future| tokio::task::spawn_blocking(|| futures::executor::block_on(future));
```

2) Because `spawn_blocking` will now block the program from exiting
until the future completes, add a drop guard to pipeline install methods
which uninstalls the global tracer provider (and in the otlp case in the
future also the meter provider) so the exporters properly shut down,
causing the future to resolve.

    Alternatives considered here could be:

    a) let users unset the global(s) themselves (but they would likely
forget, causing the same blocking behavior)
    b) return a method that could be called instead of a drop guard
(similar problems)
    c) Current impl uses custom uninstall objects per exporter to make
the struct `opentelemetry_jaeger::Uninstall` rather than a more generic
`global::TracerProviderGuard` or `global::UninstallPipeline`,
alternative name suggestions could be helpful if they improve clarity.
    c) Return a pipeline object that contains a `Tracer`, and could be
the drop guard itself, maybe cleaner API but less explicit, possibly a
better choice.

3) Because `Tracer`s currently retain strong references to their
`TracerProvider`, and shutdown is called on the inner's drop, it is
still never called. To fix this, update the `Tracer` to instead contain
a `Weak<TracerProviderInner>` reference, which can be upgraded when
needed. This does force other logic like span creation to consider the
case where the `TracerProvider` has been dropped on shutdown, but making
this case explicit seems beneficial in most cases.

4) Additional clarity/consistency tweaks:

    * Rename `global::set_provider` to `global::set_tracer_provider`
    * Rename `global::trace_provider` to `global::tracer_provider`
    * Rename `ProviderInner` to `TracerProviderInner`