Store parent context in span builder
The span builder currently holds a parent `SpanContext`, and `Context`
information is passed separately when starting a span directly or via a
builder. This poses a few problems, first you can assign a context which
contains an active span _as well as_ a builder with a parent span
context, which can lead to confusion, and secondly we have to construct
a wrapper context when sampling even in cases when the passed in context
already contains an active span.

This patch resolves this by changing the span builder's parent context
to instead store a `Context` directly. This allows the builder methods
to have a sigle context to look at when searchig for parent span
contexts, and allows the wrapper context to onlly be created for
sampling when it is facilitating a remote parent.