metrics: implement batch observer (#429)
This patch adds an implementation of the metrics batch observer. The API
is not identical to the example in [the
spec](https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/metrics/api.md#batch-observer)
as it is inconvenient to register metrics after they are moved into the
observer callback in rust. Instead the `Meter::batch_observer` method
accepts a closure in which instruments can be registered before being
moved into the callback.

```rust
meter.batch_observer(|batch| {
    let inst = batch.u64_sum_observer("example").init();

    move |result| {
        result.observe(&[KeyValue::new("a", "1")], inst.observation(42)]);
    }
});
```