feat: add tokio-rt-current-thread feature to support single thread tokio runtime use case.
To improve our performance, it would be good if we can use tokio::spawn instead of tokio::spawn_blocking.

The problem with using tokio::spawn to spawn background task that send spans is when shutting down, we need to block on the shutdown function and sending the spans at the same time. In multiple thread runtime, those tasks can be arranged in different thread. But in single runtime we will have a dead lock.

The proposed solution is to spawn the background task in a separate thread from the tokio runtime.