fix: opentelemetry-otlp will try to create a new tokio runtime if async feature is not set.
This was introduced in #467. I believe the async feature was there to spin up a tokio runtime if the exporter was not already in one. It's mostly because the tonic cannot work without a tokio runtime.

Renamed the feature to be less confusing.

Also clean up all the clippy problems in metrics module and add some links into doc.