Simple span processor perf
Currently the simple span processor his _highly_ sensitive to any
latency in its underlying exporter as it will block the current thread
for the full export duration when each span ends.

This patch addresses this by moving span exporting to a separate thread,
and communicating via channels.

Spans are still exported as soon as they end, and shutdown will wait for
all spans to be successfully exported, preserving the simple span
processor semantics of immediate exports, and never dropping data.