examples: add comment why unused variable is relevant
When checking the examples, make sure that it is explicit for readers
that it's important to bind traces and metrics initialization to
unused variables, so they are not dropped early and will live for the
whole lifetime of the containing block.

Missing to assign to an unused variable will result in an early drop
of the reporting logic, and no traces or metrics will be reported.