fix: Mapping between Jaeger processes and Otel process.  (#663)
* fix: Mapping between Jaeger processes and Otel process.

The spec maps resource tags to process tags and `service.name` entry in resource to be the service name of the process.

####  `Unknown_service` was used as `service.name` process tags even though users provided another name via `with_service_name` function.

Service name serves a special purpose for Jaeger as it requires every span to have a service name. In the open telemetry model, `service.name` is just a resource, which should be provided by users or have a default value but users can override it to be empty.

To address the difference between those two models. We need to answer the following questions.

1. Should we store the service name within the exporter or store it in resource and extract when exporting?
2. What's the priority of different ways to set the service name?
3. Should we report the `service.name` as a process tags/resource for jaeger spans?

In this PR, we implemented the following process
1. We store the service name as part of the `process` field in the exporter
2. The priority of different methods are listed below from high to low
- `with_service_name` method in `PipelineBuilder`
- passing `service.name` resource from `with_trace_config` method in `PipelineBuilder`
- SDK provided service name(it can come from env vars or default `unknown_service`)
3. We append a `service.name` process tag for each Jaeger span

#### Duplicate process tags

Process tags can be set via `with_tags` function in Jaeger pipeline or by the resource within the trace config(`with_trace_config` function in Jaeger pipeline).

We didn't de-duplicate entries from those two functions.

For this problem, we should deprecate the `with_tags` method and asks users to use `with_trace_config` and store the process tags/resource only in one place.

We can store the process tags in either of the following places
- exporter's process tags
- trace config resources

From a performance standpoint, we should store the tags in the exporter's process tags. Jaeger clients only require a process instance for one batch. If we store the tags as resources, we will copy them for each span in the batch, which will be discarded later in Jaeger clients.

However, storing tags in exporters may cause confusion when users install multiple exporters in the tracer provider. Other exporters could get the resource from the trace config while Jaeger exporter will use resources/tags stored in itself.

* fix: deprecated methods

* docs: add docs around tags, process tags and service name.

* fix: make clippy happy