opentelemetry-jaeger: Fix clearing span context in Propagator
When `Propagator::extract_with_context()` is run with an Extractor empty
of trace metadata, it returns a null (invalid) Context, clearing
TraceId, and disabling downwards propagation. As there is already an
implicit TraceId (if we're in a Span), we'd like to use it, traces have
to start somewhere.

This just aligns the mechanics with opentelemetry-sdk: if span
extraction errors out we return a clone of the existing context.