Append /v1/traces for http endpoints
We already append /v1/traces for grpc-tonic, start appending
/v1/traces to http URLs, as well as using the
OTEL_EXPORTER_OTLP_TRACES_ENDPOINT if it is set, as specified
by the standard.

To preserve backwards compatability we don't append /v1/traces
to URLs passed to the new_http function if it already contains
a path component.  Otherwise we would break existing code that
is specifying the /v1/traces, or other specific path.

Signed-off-by: Þórhallur Sverrisson <thorhallur.sverrisson@rb.is>