Fix EvictedQueue bug with zero capacity
The EvictedQueue was checking for the length _before_ inserting, and
popping extra items, then doing the insertion. In the case where the
capacity is set to zero, it caused the pop operation to be a no-op on
the first insert, and then insert an item anyway.

This commit fixes the issue by moving the length check after the insert
and popping any extra items.

Fixes #1148