Fix deriving endpoint URLs from OTEL_EXPORTER_OTLP_ENDPOINT
According to the spec, if OTEL_EXPORTER_OTLP_ENDPOINT env variable is
set, the full URL to use for traces is formed by appending
"/v1/traces" to OTEL_EXPORTER_OTLP_ENDPOINT. Similarly, the metrics
URL is formed by appending "/v1/metrics". We got that slightly wrong
in HttpExporterBuilder::build_client() and
TonicExportBuilder::build_channel(). They would append the path to the
value in the config, but if the env variable was set, they would use
it as is. Until commit 77db498ead, we were doing this correctly in
`SpanExporter::new_tonic`. Kind of. If you used `with_env` to read the
variables, it would set the ExportConfig.endpoint field, and we would
append the path to it correctly. But we got it wrong elsewhere.

As a result, if you used the HTTP protocol or the grpcio
implementation, and set
OTEL_EXPORTER_OTLP_ENDPOINT=http://foobar:4318/, the traces would be
incorrectly sent to "http://foobar:4318/", when the correct
destination would be "http://foobar:4318/v1/traces"

The timeout settings had similar issues; the per-signal TIMEOUT env
variable was not always taken into account.

To fix, move the logic for forming the full URLs into a new
resolve_endpoint() function, and use it in all the different
implementations. Similarly for the timeouts.

Add a test case with the three examples given in the OpenTelemetry
specification.