Add External Type Checking (#1603)
The purpose of this is to ensure we understand the API surface area. Anything exposed by a crate like a type can affect the possible compatibility. This means that if we bump an external type it could cause breaking changes. This tracks them to limit this possibility.

Affected Crates:

opentelemetry-otlp
opentelemetry-zipkin