feat(opentelemetry-sdk): add a constructor for SpanEvents and SpanLinks
# Changes

Add public constructors on SpanEvents and SpanLinks structs.

# Motivation

Even though all fields are public, because of the `#[non_exhaustive]` annotation on `SpanEvents`and `SpanLinks`, it's not possible to construct instances of these outside of the `opentelemetry-sdk` crate.

Since they are used in `export::SpanData`, it is problematic because it means there is no way to create span data containing any events or span links.
So there is no way to write tests on how to these fields are processed in external crate that implement the SpanExporter or SpanProcessor traits.