feat: add a readable span interface
Currently there is no way to read the span data without cloning it.
This causes performance issues when the span for span
processors that need to read the span data.

This commit introduces a new trait `ReadableSpan` in the sdk
implemented by SDK Spans that allows to read the span data without
cloning it.