fix: Ensure InstrumentationScope is properly propagated.
With the current implementation, there is no way for an InstrumentationScope
to export its version and attributes correctly.

The reason why this is happening is because when the InstrumentationScope is exported,
the target key is always `Some`. This is because when the code creates
the `scope_map`, the target key always has a value, either comming from
the log record's target, or from the name of the instrumentation scope.

That means that when the code calls `InstrumentationScope::from((&InstrumentationScope, Option<Cow<'static, str>>))`,
the option is always `Some`. This causes the new InstrumentationScope to always
from version and attributes from the incoming scope.

This change assumes that if the target and the instrumentation scope name match,
the user wants to preserve the version and attributes of the incoming scope.

There are other alternatives implementations that check whether the log record target
is None that could also be valid, but I think both would have the same outcome.